refactor(GN-16): 무한스크롤 훅 분리
