+++ src/components/MyReservationsPage/MyReservationList.tsx
@@ -1,65 +1,24 @@
 import Image from 'next/image';
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 
 import MyReservationCard from '@/components/common/ActivityCard/MyReservationCard';
 import SortDropDown from '@/components/common/Dropdown/SortDropdown';
 import { Modal } from '@/components/common/Modal';
+import useInfiniteScrollReservations from '@/hooks/useInfiniteScroll';
 import useModal from '@/hooks/useModal';
 import instance from '@/lib/apis/axios';
-import { MyReservation } from '@/types/get/reservationTypes';
 
 const ReservationList = () => {
-  const [reservations, setReservations] = useState<MyReservation[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [nextCursorId, setNextCursorId] = useState(null);
-  const [error, setError] = useState<string | null>(null);
-  const [isFirstFetch, setIsFirstFetch] = useState(true);
-  const [status, setStatus] = useState<string | null>(null);
+  const {
+    reservations,
+    loading,
+    error,
+    setError,
+    updateStatus,
+    fetchReservations,
+  } = useInfiniteScrollReservations(null);
 
   const { modalProps, openModal } = useModal();
-
-  const fetchReservations = useCallback(
-    async (reset: boolean = false) => {
-      setLoading(true);
-      try {
-        let url = `/my-reservations?size=10`;
-        if (nextCursorId && !reset) {
-          url += `&cursorId=${nextCursorId}`;
-        }
-        if (status) {
-          url += `&status=${status}`;
-        }
-        const { data } = await instance.get(url);
-        setReservations((prevReservations) =>
-          isFirstFetch || reset
-            ? data.reservations
-            : [...prevReservations, ...data.reservations],
-        );
-        setNextCursorId(data.cursorId);
-        setIsFirstFetch(false);
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError(String(err));
-        }
-      } finally {
-        setLoading(false);
-      }
-    },
-    [nextCursorId, status],
-  );
-
-  const handleScroll = useCallback(() => {
-    const scrollPosition =
-      window.innerHeight + document.documentElement.scrollTop;
-    const threshold = document.documentElement.offsetHeight - 100;
-
-    if (scrollPosition < threshold || loading || nextCursorId === null) {
-      return;
-    }
-    fetchReservations();
-  }, [loading, nextCursorId, fetchReservations]);
 
   const handleFilterSelect = (option: string) => {
     const statusMapping: { [key: string]: string } = {
@@ -70,9 +29,7 @@ const ReservationList = () => {
       '체험 완료': 'completed',
     };
     const selectedStatus = statusMapping[option] || null;
-    setNextCursorId(null);
-    setIsFirstFetch(true);
-    setStatus(selectedStatus);
+    updateStatus(selectedStatus);
   };
 
   const handleCancelReservation = async (reservationId: number) => {
@@ -82,7 +39,6 @@ const ReservationList = () => {
 
       await fetchReservations(true);
     } catch (error) {
-      console.error(`예약 ${reservationId} 취소 중 오류 발생:`, error);
       setError('예약 취소 중 오류가 발생했습니다.');
     }
   };
@@ -92,15 +48,6 @@ const ReservationList = () => {
       onConfirm: () => handleCancelReservation(reservationId),
     });
   };
-
-  useEffect(() => {
-    fetchReservations();
-  }, [status]);
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, [handleScroll]);
 
   if (error) return <div>Error: {error}</div>;
 
@@ -129,7 +76,7 @@ const ReservationList = () => {
             height={160}
             className="mb-4"
           />
-          <p className="text-gray-500">아직 체험이 없어요</p>
+          <p className="text-kv-gray-500 kv-text-xl">아직 체험이 없어요</p>
         </div>
       ) : (
         <div>
